processing/process_gazettes/requirements.txt
kafka-python==1.4.7
